[IMP] Co-SLAM: add modifications to use the algorithm without cuda
Changelog:
- Add submodule hash-grid-encoding that has encoding implementations
  withoud tynicudann.
- Add submodule pytorch3d because it can not be installed by pip or
  conda.
- Replace encodings with the encodings in the new submodule.
- Use float method to convert to float32 the inputs to the networks in
  order to match the type of the weight matrix [1]. This solves the
  error: RuntimeError: expected m1 and m2 to have the same dtype, but
  got: double != float

[1] https://discuss.pytorch.org/t/runtimeerror-mat1-and-mat2-must-have-the-same-dtype/166759/2

+++ model/scene_rep.py
@@ -7,6 +7,7 @@ from .encodings import get_encoder
 from .decoder import ColorSDFNet, ColorSDFNet_v2
 from .utils import sample_pdf, batchify, get_sdf_loss, mse2psnr, compute_loss
 
+
 class JointEncoding(nn.Module):
     def __init__(self, config, bound_box):
         super(JointEncoding, self).__init__()
@@ -148,12 +149,12 @@ class JointEncoding(nn.Module):
         Returns:
             raw: [N_rays, N_samples, 4]
         '''
-        inputs_flat = torch.reshape(query_points, [-1, query_points.shape[-1]])
+        inputs_flat = torch.reshape(query_points, [-1, query_points.shape[-1]]).to(torch.float64)
 
-        embed = self.embed_fn(inputs_flat)
-        embe_pos = self.embedpos_fn(inputs_flat)
+        embed = self.embed_fn(inputs_flat).float()
+        embe_pos = self.embedpos_fn(inputs_flat).float()
         if not self.config['grid']['oneGrid']:
-            embed_color = self.embed_fn_color(inputs_flat)
+            embed_color = self.embed_fn_color(inputs_flat).float()
             return self.decoder(embed, embe_pos, embed_color)
         return self.decoder(embed, embe_pos)
